+++ camber/store/modelstore.py
@@ -45,6 +45,7 @@ _MODEL_TYPES = {
     "pump_flow": LoadBaseline,
     "pump_power": LoadBaseline,
     "fan_efficiency": LoadBaseline,
+    "filter_loading": LoadBaseline,
     "pump_head": LoadBaseline,
     "loop_deltat": LoadBaseline,
     "loop_dp": LoadBaseline,